voice & keyboard move fully disentangled
also bugfixes and square selection
